refactor(nifs): err in `nifs::FoldingScheme`.
**Motivation**
The errors will be different for vanila & protogalaxy, so a slightly
different design of `trait FoldingScheme` methods is needed
Part of #266

**Overview**
Old generic error type moved to `nifs::vanila` module
